backend/requirements.txt
fastapi
uvicorn[standard]
numpy
torch
torchvision
pillow
scikit-learn
python-multipart